Attempt to improve the keyboard input handling
Attempted to improve some of the handling of keyboard input events. There are
some huge problems with repeat events of non-printable keys in SDL,
especially backspace, return, tab, and the arrow keys. The repeat rate is not
as one would expect, and alt+tab is a problem. These are problems are now
noted in the comments.